trivial new object to provide a rectangle with clear interiors that still
2002-01-21  Jody Goldberg <jody@gnome.org>

	* src/item-acetate.c : trivial new object to provide a rectangle with
	  clear interiors that still collects events.

	* src/sheet-control-gui.c (set_item_x_y) : optionally hide a point.
	(scg_object_update_bbox) : hide middle control points when the bounds
	  are too close together.
	(set_acetate_coords) : use new simpler item-acetate, and provide a
	  cute stipple to make things clearer.

+++ src/item-cursor.c
@@ -540,13 +540,13 @@ item_cursor_point (GnomeCanvasItem *item, double x, double y, int cx, int cy,
 	*actual_item = NULL;
 
 	if (cx < item->x1-3)
-		return INT_MAX;
+		return DBL_MAX;
 	if (cx > item->x2+3)
-		return INT_MAX;
+		return DBL_MAX;
 	if (cy < item->y1-3)
-		return INT_MAX;
+		return DBL_MAX;
 	if (cy > item->y2+3)
-		return INT_MAX;
+		return DBL_MAX;
 
 	/* FIXME: the drag handle for ITEM_CURSOR_SELECTION needs work */
 	if ((cx < (item->x1 + 4)) || (cx > (item->x2 - 8)) ||
